pyproject.toml
[project]
name = "ai_systems_root"
version = "0.1.0"
description = "Shared dependencies and tooling for the ai-systems monorepo."
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "black>=25.11.0",
    "loguru>=0.7.3",
    "pre-commit>=4.5.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "ruff>=0.14.7",
]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312", "py313"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
exclude = [
  ".git",
  "__pypackages__",
  ".venv",
  "venv",
  "build",
  "dist",
  "legacy",
]
fix = true

[tool.ruff.lint]
select = ["E", "F", "W", "C", "B", "I", "R", "T"]
ignore = []